+++ Backend/eewms/src/main/java/com/eewms/services/IInventoryService.java
@@ -5,6 +5,7 @@ import com.eewms.dto.inventory.WarehouseStockRowDTO;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,5 +29,5 @@ public interface IInventoryService {
     /**
      * Số lượng on-hand của một sản phẩm tại một kho (Optional).
      */
-    Optional<Integer> getOnHand(Long productId, Integer warehouseId);
+    Optional<BigDecimal> getOnHand(Long productId, Integer warehouseId);
 }
